Fix an issue with queue depth / time estimates
If two jobs were started at the same time and both required the same anime:
Job A adds 1 to the queue length (now 1)
Job A adds the pending anime
Job A lastQueuePosition is 1
Job B adds 1 to the queue length (now 2)
Job B attempts to add the pending anime, but it fails due to the write condition
Job B adds 1 to the processed items (now 1)
Job B lastQueuePosition is 1 (based on the returned value from the failed write)

Job A & B are both estimated to take 0 seconds, since the queue length and processed jobs match.

I have switched this so that if we hit a write conflict, we decrement the queue length instead. This does mean that potentially the queue positions assigned to jobs might not be in order, but hopefully it happens rarely enough to not be a big problem. The estimates are not intended to be 100% accurate anyway.