Split into three Lambda functions
The "limited" Lambda function has a concurrency limit of 1. This is to ensure we don't hit Jikan too hard.

The "lightweight" Lambda function has 128MB of memory. This can be used for compute-light tasks, such as checking the status of a job.

The "heavyweight" Lambda function has 768MB of memory. This greatly improves the speed of job initialisation (44 seconds -> 8 seconds). In future I'd like to use it for the crunching part too.

This change is especially important because it brings us under the 30 second API Gateway timeout. I'm a bit nervous that large MAL profiles might still exceed the limit, but this is a big improvement at least.